+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - If a module is not listed in ``--ftp-requirements-module-extra-mapping``, FTP041 will
   now only allow install requirements instead of all requirements
 - FTP108 now ignores usages of `NoReturn` in `Callable` like `x = Callable[[int], NoReturn]`
+- FTP139 which can find uses of `profile` and `cProfile` which should be replaced with
+  `profiling.tracing` within python 3.15 onwards.
 
 ## 26.1.28.0
 - Support recursive optional requirements in FTP041